agent/requirements.txt
# LiveKit Agents with STT-LLM-TTS pipeline (cloud-based)
livekit-agents[deepgram,groq,cartesia]~=1.2
